requirements.txt
catboost==1.0.6
fastapi==0.83.0
numpy==1.21.5
pandas==1.4.3
pydantic==1.10.2
pytest==7.1.2
scikit_learn==1.1.1
pandas-profiling==3.2.0
jupyterlab==3.4.4
ipywidgets==7.6.5
flake8==4.0.1
uvicorn==0.18.3
requests==2.28.1
hydra-core==1.2.0
